backend/requirements.txt
django-cors-headers==4.7.0
django-axes==6.0.0
Django==5.1.7
djangorestframework==3.15.2
djangorestframework-simplejwt==5.5.0
numpy==2.2.4
scikit-learn==1.6.1
psycopg2-binary==2.9.10  
openpyxl==3.1.5
pandas==2.2.3
dash==3.0.1
reportlab==4.3.1
bleach==6.2.0
flask-cors==5.0.1
django-extensions==3.2.3
matplotlib==3.10.1
cryptography==44.0.2
drf-yasg==1.21.7